containers: Start http server in config script
No need for that separate script that the user would need to run.

Added "hostname:" in docker-compose, so that http-servers
can respond with their names. This helps when experimenting
with ssh-tunnels.

It also gives a nice prompt when the user logs in to
the various containers. For example:

root@ssh-client:/#